requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
plotly.express